change to float32

+++ tinygrad/function.py
@@ -45,8 +45,8 @@ class Sin(Function):
     # q = x.e(BinaryOps.DIV, x.const(2 * math.pi))
     # q = q.cast(dtypes.int32).cast(old_dtype)
     # x = x.e(BinaryOps.SUB, x.e(BinaryOps.DIV, x.const(2 * math.pi)).cast(dtypes.int32).cast(old_dtype).e(BinaryOps.MUL, x.const(2 * math.pi)))
-    x = x.cast(dtypes.float64)
-    x = x.e(BinaryOps.SUB, x.e(BinaryOps.DIV, x.const(2 * math.pi)).cast(dtypes.int32).cast(dtypes.float64).e(BinaryOps.MUL, x.const(2 * math.pi)))
+    x = x.cast(dtypes.float32)
+    x = x.e(BinaryOps.SUB, x.e(BinaryOps.DIV, x.const(2 * math.pi)).cast(dtypes.int32).cast(dtypes.float32).e(BinaryOps.MUL, x.const(2 * math.pi)))
     # x = x.e(BinaryOps.SUB, q.e(BinaryOps.MUL, x.const(2 * math.pi)))
 
     # no_terms = 20
